got rid of store directory and moved actions and reducers folders to src folder. changed direction of action imports in reducer.js
