+++ app/src/reducers/index.js
@@ -1,6 +1,6 @@
 import { combineReducers} from 'redux';
 import { reducer } from './reducer';
 
-export default combineReducers({
-    reducer
+export const rootReducer = combineReducers({
+    reducer,
 })
